requirements.txt
# Runtime
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4

# Database
asyncpg==0.29.0

# AWS
boto3==1.34.69

# Geospatial
GDAL==3.7.3
rasterio==1.3.10
Shapely==2.0.4
pyproj==3.6.1
numpy==1.26.4

# ML
torch==2.2.2
torchvision==0.17.2

# LLM
anthropic==0.25.0

# Image
Pillow==10.3.0

# Graph
networkx==3.3

# Training (not used in prod — harmless to include)
tensorboard==2.16.2
albumentations==1.4.3
